Project now works with cocoapods.
The workspace in the project folder is able to build OSX framework, iOS static lib. The static iOS framework still needs a bit of work to link together properly.

The OSX target has been consolidated into the same project. This project references the same files as the iOS version.

The sample Simpletodo project is set up as a local pod. Thus, changing files here will automatically change the files in the pod. This should make it easy to work simultaneously on the framework as well as another project.

I've tested the framework